+++ db/migrations/install/000000-000001-seed-db.sql
@@ -43,4 +43,5 @@ INSERT INTO migrations (migration, applied) VALUES
 	('000000-000019-populate-asset-catalog[pgsql].php', now()),
 	('000000-000020-richtext-html-to-json[pgsql].php', now()),
 	('000000-000021-create-reference-indexes[pgsql].php', now()),
-	('000000-000022-materialize-node-titles[pgsql].sql', now());
+	('000000-000022-materialize-node-titles[pgsql].sql', now()),
+	('000000-000023-user-panel-locale.sql', now());