+++ db/sql/users/savePanelLocale.sql
@@ -1,0 +1,6 @@
+UPDATE /*:cms.prefix:*/users
+SET
+	panel_locale = :locale,
+	editor = :usr
+WHERE
+	usr = :usr;